[#21] fix: 리뷰 후 1차 수정 사항
- InstituteDetailResponse로 명명 수정
- 위도, 경도 BigDecimal로 변경

+++ src/main/java/com/gamsa/activity/domain/Institute.java
@@ -1,5 +1,6 @@
 package com.gamsa.activity.domain;
 
+import java.math.BigDecimal;
 import lombok.Builder;
 import lombok.Getter;
 
@@ -10,8 +11,8 @@ public class Institute {
     private Long instituteId;
     private String name;
     private String location;
-    private long latitude;
-    private long longitude;
+    private BigDecimal latitude;
+    private BigDecimal longitude;
     // Todo 시군구 코드
     private String phone;
 }
